force int64 to make tests more stable

+++ v1/tv/internal/traceview/metrics_linux.go
@@ -35,7 +35,7 @@ func addHostMetrics(bbuf *bsonBuffer, index *int) {
 		memTotal := strings.Fields(s) // MemTotal: 7657668 kB
 		if len(memTotal) == 3 {
 			if total, err := strconv.Atoi(memTotal[1]); err == nil {
-				addMetricsValue(bbuf, index, "TotalRAM", total*1024)
+				addMetricsValue(bbuf, index, "TotalRAM", int64(total*1024))
 			}
 		}
 	}
@@ -45,7 +45,7 @@ func addHostMetrics(bbuf *bsonBuffer, index *int) {
 		memFree := strings.Fields(s) // MemFree: 161396 kB
 		if len(memFree) == 3 {
 			if free, err := strconv.Atoi(memFree[1]); err == nil {
-				addMetricsValue(bbuf, index, "FreeRAM", free*1024) // bytes
+				addMetricsValue(bbuf, index, "FreeRAM", int64(free*1024)) // bytes
 			}
 		}
 	}
